Rename directory cuda_hip to cuda-hip to fix amalgamation
