+++ .clang-format
@@ -135,20 +135,20 @@ UseTab: Never
 #IfMacros: []
 IncludeCategories:
   # Local headers (in "") above all else
-  - Regex: '"([A-Za-z0-9.\/-_])+"'
+  - Regex: '"([A-Za-z0-9./_\-])+"'
     Priority: 1
   # "alpaka/foo.hpp" after local headers (occur inside alpaka)
-  - Regex: '"alpaka/([A-Za-z0-9.\/-_])+"'
+  - Regex: '"alpaka/([A-Za-z0-9./_\-])+"'
     Priority: 2
   # <alpaka/foo.hpp> after local headers (occur outside alpaka in examples and test)
-  - Regex: '<alpaka/([A-Za-z0-9.\/-_])+>'
+  - Regex: '<alpaka/([A-Za-z0-9./_\-])+>'
     Priority: 3
   # C++ standard library headers are the last group to be included
-  - Regex: '<([A-Za-z0-9\/-_])+>'
+  - Regex: '<([A-Za-z0-9/_\-])+>'
     Priority: 5
   # Includes that made it this far are third-party headers and will be placed
-  # below alpaka's includes
-  - Regex: '<([A-Za-z0-9.\/-_])+>'
+  # below alpaka's includes 
+  - Regex: '<([A-Za-z0-9./_\-])+>'
     Priority: 4
 # Macros: []
 # NamespaceMacros: []
