+++ include/alpaka/queue/QueueUniformCudaHipRtBlocking.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "alpaka/queue/cuda_hip/QueueUniformCudaHipRt.hpp"
+#include "alpaka/queue/cuda-hip/QueueUniformCudaHipRt.hpp"
 
 #if defined(ALPAKA_ACC_GPU_CUDA_ENABLED) || defined(ALPAKA_ACC_GPU_HIP_ENABLED)
